+++ cpp/libjoynrclustercontroller/ClusterControllerSettings.cpp
@@ -98,10 +98,6 @@ void ClusterControllerSettings::checkSettings()
         setMqttMulticastTopicPrefix(DEFAULT_MQTT_UNICAST_TOPIC_PREFIX());
     }
 
-    if (!_settings.contains(SETTING_USE_ONLY_LDAS())) {
-        setUseOnlyLDAS(DEFAULT_USE_ONLY_LDAS());
-    }
-
     if (!_settings.contains(SETTING_PURGE_EXPIRED_DISCOVERY_ENTRIES_INTERVAL_MS())) {
         setPurgeExpiredDiscoveryEntriesIntervalMs(
                 DEFAULT_PURGE_EXPIRED_DISCOVERY_ENTRIES_INTERVAL_MS());
@@ -163,25 +159,6 @@ void ClusterControllerSettings::checkSettings()
 
     if (!_settings.contains(SETTING_ACCESS_CONTROL_ENABLE())) {
         setEnableAccessController(DEFAULT_ENABLE_ACCESS_CONTROLLER());
-    } else if (enableAccessController() && !getUseOnlyLDAS()) {
-        assert(_settings.contains(
-                SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_ADDRESS()));
-        assert(_settings.contains(
-                SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_PARTICIPANTID()));
-
-        if (!_settings.contains(SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_ADDRESS())) {
-            JOYNR_LOG_ERROR(logger(),
-                            "Configuration error. Access controller is enabled but "
-                            "no {} was defined.",
-                            SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_ADDRESS());
-        }
-        if (!_settings.contains(
-                    SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_PARTICIPANTID())) {
-            JOYNR_LOG_ERROR(logger(),
-                            "Configuration error. Access controller is enabled but "
-                            "no {} was defined.",
-                            SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_PARTICIPANTID());
-        }
     }
 
     if (!_settings.contains(SETTING_ROUTED_MESSAGE_PRINT_INTERVAL_S())) {
@@ -260,12 +237,6 @@ const std::string& ClusterControllerSettings::SETTING_UDS_ENABLED()
     return value;
 }
 
-const std::string& ClusterControllerSettings::SETTING_USE_ONLY_LDAS()
-{
-    static const std::string value("access-control/use-ldas-only");
-    return value;
-}
-
 const std::string& ClusterControllerSettings::SETTING_ACCESS_CONTROL_ENABLE()
 {
     static const std::string value("access-control/enable");
@@ -275,20 +246,6 @@ const std::string& ClusterControllerSettings::SETTING_ACCESS_CONTROL_ENABLE()
 const std::string& ClusterControllerSettings::SETTING_ACCESS_CONTROL_AUDIT()
 {
     static const std::string value("access-control/audit");
-    return value;
-}
-
-const std::string& ClusterControllerSettings::
-        SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_ADDRESS()
-{
-    static const std::string value("access-control/global-domain-access-controller-address");
-    return value;
-}
-
-const std::string& ClusterControllerSettings::
-        SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_PARTICIPANTID()
-{
-    static const std::string value("access-control/global-domain-access-controller-participantid");
     return value;
 }
 
@@ -487,11 +444,6 @@ bool ClusterControllerSettings::DEFAULT_MULTICAST_RECEIVER_DIRECTORY_PERSISTENCY
 bool ClusterControllerSettings::DEFAULT_ENABLE_ACCESS_CONTROLLER()
 {
     return false;
-}
-
-bool ClusterControllerSettings::DEFAULT_USE_ONLY_LDAS()
-{
-    return true;
 }
 
 const std::string& ClusterControllerSettings::DEFAULT_CLUSTERCONTROLLER_SETTINGS_FILENAME()
@@ -890,28 +842,6 @@ void ClusterControllerSettings::setAclAudit(bool audit)
     _settings.set(SETTING_ACCESS_CONTROL_AUDIT(), audit);
 }
 
-std::string ClusterControllerSettings::getGlobalDomainAccessControlAddress() const
-{
-    return _settings.get<std::string>(
-            SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_ADDRESS());
-}
-
-std::string ClusterControllerSettings::getGlobalDomainAccessControlParticipantId() const
-{
-    return _settings.get<std::string>(
-            SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_PARTICIPANTID());
-}
-
-bool ClusterControllerSettings::getUseOnlyLDAS() const
-{
-    return _settings.get<bool>(SETTING_USE_ONLY_LDAS());
-}
-
-void ClusterControllerSettings::setUseOnlyLDAS(bool useOnlyLDAS)
-{
-    _settings.set(SETTING_USE_ONLY_LDAS(), useOnlyLDAS);
-}
-
 std::string ClusterControllerSettings::getLocalCapabilitiesDirectoryPersistenceFilename() const
 {
     return _settings.get<std::string>(SETTING_LOCAL_CAPABILITIES_DIRECTORY_PERSISTENCE_FILENAME());
@@ -1076,8 +1006,6 @@ void ClusterControllerSettings::printSettings() const
                    SETTING_LOCAL_DOMAIN_ACCESS_STORE_PERSISTENCE_FILENAME(),
                    getLocalDomainAccessStorePersistenceFilename());
 
-    JOYNR_LOG_INFO(logger(), "SETTING: {} = {}", SETTING_USE_ONLY_LDAS(), getUseOnlyLDAS());
-
     JOYNR_LOG_INFO(logger(),
                    "SETTING: {} = {}",
                    SETTING_ACCESS_CONTROL_ENABLE(),
@@ -1099,20 +1027,6 @@ void ClusterControllerSettings::printSettings() const
                    "SETTING: {} = {}",
                    SETTING_PURGE_EXPIRED_DISCOVERY_ENTRIES_INTERVAL_MS(),
                    getPurgeExpiredDiscoveryEntriesIntervalMs());
-
-    if (_settings.get<bool>(SETTING_ACCESS_CONTROL_ENABLE())) {
-        JOYNR_LOG_INFO(logger(),
-                       "SETTING: {} = {}",
-                       SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_ADDRESS(),
-                       _settings.get<std::string>(
-                               SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_ADDRESS()));
-        JOYNR_LOG_INFO(
-                logger(),
-                "SETTING: {} = {}",
-                SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_PARTICIPANTID(),
-                _settings.get<std::string>(
-                        SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_PARTICIPANTID()));
-    }
 }
 
 } // namespace joynr