+++ cpp/libjoynrclustercontroller/include/joynr/ClusterControllerSettings.h
@@ -60,15 +60,11 @@ public:
     static const std::string& SETTING_PURGE_EXPIRED_DISCOVERY_ENTRIES_INTERVAL_MS();
     static const std::string& SETTING_WS_TLS_PORT();
     static const std::string& SETTING_WS_PORT();
-    static const std::string& SETTING_USE_ONLY_LDAS();
     static const std::string& SETTING_ACCESS_CONTROL_AUDIT();
     static const std::string& SETTING_UDS_ENABLED();
     static const std::string& SETTING_WEBSOCKET_ENABLED();
 
     static const std::string& SETTING_ACCESS_CONTROL_ENABLE();
-    static const std::string& SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_ADDRESS();
-    static const std::string&
-    SETTING_ACCESS_CONTROL_GLOBAL_DOMAIN_ACCESS_CONTROLLER_PARTICIPANTID();
     static const std::string& SETTING_ACL_ENTRIES_DIRECTORY();
     static const std::string& SETTING_GLOBAL_CAPABILITIES_DIRECTORY_COMPRESSED_MESSAGES_ENABLED();
     static const std::string& SETTING_ROUTED_MESSAGE_PRINT_INTERVAL_S();
@@ -88,7 +84,6 @@ public:
     static bool DEFAULT_MULTICAST_RECEIVER_DIRECTORY_PERSISTENCY_ENABLED();
     static int DEFAULT_PURGE_EXPIRED_DISCOVERY_ENTRIES_INTERVAL_MS();
     static bool DEFAULT_ENABLE_ACCESS_CONTROLLER();
-    static bool DEFAULT_USE_ONLY_LDAS();
     static bool DEFAULT_ACCESS_CONTROL_AUDIT();
     static std::uint64_t DEFAULT_MESSAGE_QUEUE_LIMIT();
     static std::uint64_t DEFAULT_PER_PARTICIPANTID_MESSAGE_QUEUE_LIMIT();
@@ -179,12 +174,6 @@ public:
 
     bool aclAudit() const;
     void setAclAudit(bool enable);
-
-    std::string getGlobalDomainAccessControlAddress() const;
-    std::string getGlobalDomainAccessControlParticipantId() const;
-
-    bool getUseOnlyLDAS() const;
-    void setUseOnlyLDAS(bool useLDASonly);
 
     std::string getLocalCapabilitiesDirectoryPersistenceFilename() const;
     void setLocalCapabilitiesDirectoryPersistenceFilename(const std::string& filename);